refactor(runtime): extract RuntimeCore into runtime/core.rs
`ApplicationState` was a misleading name: in a TEA framework "application
state" reads as the user's model, but this type is the runtime's own
execution resources for one app run — the app instance plus the
message/quit channels, input mux, subscription manager, and command-task
set, along with their lifecycle operations. Rename it to `RuntimeCore`
and move it (with its unit tests) out of the ~1.6k-line runtime.rs into a
dedicated `runtime::core` module, matching the file=type snake_case
convention (app_input.rs/AppInputs, frame_scheduler.rs/FrameScheduler).

- `Runtime` now holds a `core: RuntimeCore<App>` and delegates.
- Fields/methods are `pub(super)`. This is deliberate, not a leak: the
  `run()` loop's `select!` borrows `app_inputs` and `quit_rx` in separate
  branches, requiring disjoint `&mut` borrows of individual fields;
  wrapping them in whole-`self` methods would hold `&mut self` across the
  whole `select!` and conflict with the branch handlers. The doc comment
  records this.
- `use self::core::RuntimeCore` disambiguates the submodule from the
  built-in `core` crate (avoids E0659).

The shared `TestApp`/`TestMessage` fixture is used by tests on both sides
of the split, so it moves to the existing crate-wide `test_support`
module. Fixtures used only by the moved tests (`AppWithInitCommand`,
`AppWithStringFlags`) move with them; `EventCounter`, `RedrawControlApp`,
and `SubCountingApp` stay with their runtime-level tests.

runtime.rs drops ~515 lines. No behavior change.
