+++ src/lib.rs
@@ -114,6 +114,12 @@ mod test_support {
 
     use std::sync::Mutex;
 
+    use ratatui::Frame;
+
+    use crate::application::Application;
+    use crate::command::{Action, Command};
+    use crate::subscription::Subscription;
+
     /// Serializes tests that install a process-global panic hook or deliberately
     /// trigger panics.
     ///
@@ -121,4 +127,56 @@ mod test_support {
     /// records hook activity and a test that panics must not run concurrently, or
     /// the panicking test's hook invocation would pollute the recording one.
     pub static PANIC_HOOK_GUARD: Mutex<()> = Mutex::new(());
+
+    /// A minimal counter [`Application`] shared by the runtime and runtime-core
+    /// tests. Increments on [`TestMessage::Increment`] and quits on
+    /// [`TestMessage::Quit`].
+    #[derive(Debug)]
+    pub struct TestApp {
+        pub counter: i32,
+        should_quit: bool,
+    }
+
+    #[derive(Debug, Clone)]
+    #[allow(dead_code)]
+    pub enum TestMessage {
+        Increment,
+        Quit,
+    }
+
+    impl Application for TestApp {
+        type Message = TestMessage;
+        type Flags = i32;
+
+        fn new(initial: i32) -> (Self, Command<Self::Message>) {
+            (
+                Self {
+                    counter: initial,
+                    should_quit: false,
+                },
+                Command::none(),
+            )
+        }
+
+        fn update(&mut self, msg: Self::Message) -> Command<Self::Message> {
+            match msg {
+                TestMessage::Increment => {
+                    self.counter += 1;
+                    Command::none()
+                }
+                TestMessage::Quit => {
+                    self.should_quit = true;
+                    Command::effect(Action::Quit)
+                }
+            }
+        }
+
+        fn view(&self, _frame: &mut Frame<'_>) {
+            // No-op for testing
+        }
+
+        fn subscriptions(&self) -> Vec<Subscription<Self::Message>> {
+            vec![]
+        }
+    }
 }
